requirements.txt
git+https://gitlab.com/flywheel-io/scientific-solutions/gears/deid-export
flywheel-sdk
